[개발] (메뉴) ENTITY_LOG, ENTITY_HISTORY에 update, delete , insert 쌓이도록 메뉴 코드 변경

+++ src/main/java/com/dejay/framework/vo/search/menu/MenuSearchVO.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 public class MenuSearchVO extends SearchVO {
     // 메뉴 Seq
-    private int menuSeq;
+    private Long menuSeq;
     // 메뉴 ID
     private String menuId;
